[~TASK] CouchDB: Do major code cleanup
This commit fixes a lot of documentation and CGL issues, removes use of
the shut-up operator (@), introduces more object manager use, renames a
bunch of methods and variables and tweaks some method signatures. On top
of all that the exceptions in the Client namespace are now used
correctly and a lot of type hints have been added.

Change-Id: I3ce8abece5e2335c79701fe633de120d185ba138