[+FEATURE] CouchDB Lucene implementation
CouchDB has only minimal query possibilities without like constraints
and logical or. CouchDB Lucene is a transparent index behind CouchDB
that offers all Lucene query features through a server running as
an external command.

This change adds a LuceneIndex and specialized LuceneQuery and
LuceneQueryFactory to be used in Repositories. The CouchDbBackend can
handle the special query and queries the index instead of the view.
Indexes are stored in separate design documents and stored on first
request.

For now the code is only tested with functional tests.

Change-Id: I9083a901eb717fdfa5b98813c1406227849d6678