[~TASK] Fix problem with eager loading and bidirectional associations
The problem needs to be addressed in the DataMapper, too. It relies
on the object data being fetched recursively in depth first order. The
CouchDB eager loading fetches related objects in breadth first order in
an eager way. As a result some objects are not available in the session
during reconstitution and have incomplete object data (only identifier).

Change-Id: Ib0f34698406fcf70a353292c1c8151004f399449