[BUGFIX] Fix certain side-effects caused by missing validation
The CouchDB backend relied on some side-effects caused by the validation
of any property that initialized lazy references. This change introduces
a lot of functional tests for these cases and implements a different
strategy for deleting orphaned entities. Additionally a safe removal
of entities is also possible without a revision (e.g. for lazy
objects) by fetching the document information (HEAD) prior to
deleting the document.

Change-Id: I30442503c1a46f0e2603a8f106490a7bcd7511dd