[FEATURE] Add remove strategies to keep deleted documents
If special identifiers are used it might be desirable to keep
deleted documents as empty documents with a "deleted" flag to
block the document id. This is implemented with a new
@RemoveStrategy annotation and the strategy KEEP_ID.Also an
undelete might be possible with the remove strategy SOFT_DELETE.
Note however that SOFT_DELETE will not yet filter out deleted
documents from queries automatically.

Change-Id: Id8a54cdd16dda1ac2ef3a41922bf63b48fb4a3b1